+++ molo/core/templatetags/core_tags.py
@@ -14,7 +14,7 @@ from molo.core.models import (
     Page, ArticlePage, SectionPage, SiteSettings, Languages, Tag,
     ArticlePageTags, SectionIndexPage, ReactionQuestion,
     ReactionQuestionChoice, BannerPage, get_translation_for,
-    ArticleOrderingChoices
+    ArticleOrderingChoices, ReactionQuestionResponse
 )
 
 
@@ -183,12 +183,21 @@ def bannerpages(context, position=-1):
                 'bannerpages': [banners[position]],
                 'request': context['request'],
                 'locale_code': locale,
+                'is_via_freebasics':
+                    'Internet.org' in request.META.get('HTTP_VIA', '') or
+                    'InternetOrgApp' in request.META.get(
+                        'HTTP_USER_AGENT', '') or
+                    'true' in request.META.get('HTTP_X_IORG_FBS', '')
             }
         return None
     return {
         'bannerpages': get_pages(context, pages, locale),
         'request': context['request'],
         'locale_code': locale,
+        'is_via_freebasics':
+            'Internet.org' in request.META.get('HTTP_VIA', '') or
+            'InternetOrgApp' in request.META.get('HTTP_USER_AGENT', '') or
+            'true' in request.META.get('HTTP_X_IORG_FBS', '')
     }
 
 
@@ -611,6 +620,14 @@ def load_choices_for_reaction_question(context, question):
             question).filter(language__is_main_language=True)
         return get_pages(context, choices, locale)
     return []
+
+
+@register.simple_tag()
+def load_reaction_choice_submission_count(choice, article, question):
+    if choice and article:
+        choice = choice.specific.get_main_language_page().specific
+        return ReactionQuestionResponse.objects.filter(
+            article=article, choice=choice, question=question).count()
 
 
 @register.simple_tag(takes_context=True)